+++ database/migrations/2014_10_11_000000_create_users_table.php
@@ -14,16 +14,16 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('us_username')->unique();
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
 
-            $table->integer('person_id')->unsigned()->nullable();
+            $table->bigInteger('person_id')->unsigned()->nullable();
             $table->foreign('person_id')->references('id')->on('persons');
 
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->rememberToken();
