+++ database/migrations/2021_08_31_113641_create_curriculums_table.php
@@ -14,33 +14,33 @@ class CreateCurriculumsTable extends Migration
     public function up()
     {
         Schema::create('curriculums', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('mes_name', 255)->nullable();
             $table->string('mes_res_cas', 255)->nullable();
             $table->string('mes_res_ocas')->nullable();
             $table->string('mes_cod_career')->nullable();
             $table->string('mes_title')->nullable();
             $table->string('mes_itinerary')->nullable();
-            $table->integer('mes_number_matter')->default(0);
-            $table->integer('mes_number_period')->nullable();
-            $table->integer('mes_quantity_external_matter_homologate')->nullable();
-            $table->integer('mes_quantity_internal_matter_homologate')->nullable();
+            $table->bigInteger('mes_number_matter')->default(0);
+            $table->bigInteger('mes_number_period')->nullable();
+            $table->bigInteger('mes_quantity_external_matter_homologate')->nullable();
+            $table->bigInteger('mes_quantity_internal_matter_homologate')->nullable();
             $table->date('mes_creation_date')->nullable();
             $table->string('mes_acronym', 3)->nullable();
-            $table->integer('anio')->nullable();
+            $table->bigInteger('anio')->nullable();
             $table->string('mes_description', 255)->nullable();
 
 
-            $table->integer('mes_modality_id')->unsigned();
+            $table->bigInteger('mes_modality_id')->unsigned();
             $table->foreign('mes_modality_id')->references('id')->on('catalogs');
 
-            $table->integer('type_calification_id')->unsigned();
+            $table->bigInteger('type_calification_id')->unsigned();
             $table->foreign('type_calification_id')->references('id')->on('type_califications');
 
-            $table->integer('level_edu_id')->unsigned();
+            $table->bigInteger('level_edu_id')->unsigned();
             $table->foreign('level_edu_id')->references('id')->on('education_levels');
 
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->timestamps();
