+++ database/migrations/2021_08_31_130043_create_type_students_table.php
@@ -14,15 +14,15 @@ class CreateTypeStudentsTable extends Migration
     public function up()
     {
         Schema::create('type_students', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('te_name', 255)->nullable();
             $table->string('te_description', 255)->nullable();
 
-            // $table->integer('offer_id')->unsigned();
+            // $table->bigInteger('offer_id')->unsigned();
             // $table->foreign('offer_id')->references('id')->on('offers');
            
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->timestamps();
